Fix point_search, Add specs, Update README.md
Change point_search such that it does not assume the interval
has the `include?` method. This allows for custom objects to
be used as intervals.

Add specs for the behavior of the tree with custom
interval objects. Specs include:
  * Tree creation
  * Searching for custom objects
  * Searching with custom objects

Update the readme to include examples with custom objects, and
update the notes to reflect how custom objects behave.